requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
idna==3.10
pillow==11.3.0
pypdf==5.8.0
pyshorteners==1.0.1
qrcode==8.2
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
